# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Ouderenalarm API'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/user/caregivers/update:
    post:
      summary: 'Update caregiver priorities.'
      operationId: updateCaregiverPriorities
      description: ''
      parameters: []
      responses: {  }
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caregivers:
                  type: array
                  description: ''
                  example:
                    - quis
                  items:
                    type: string
              required:
                - caregivers
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []